pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "energy-mpc-toy"
version = "0.1.0"
description = "Toy model investigating energy and MPC for retrieval" 
requires-python = ">=3.8"

dependencies = [
    "accelerate>=0.25.0",
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "einops>=0.6.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "pytest>=7.0.0",
    "datasets>=2.12.0",
    "wandb>=0.15.0",
    "matplotlib>=3.7.0",
    "psutil>=5.9.0",
    "tqdm>=4.65.0",
    "gputil>=1.4.0",
    "faiss-cpu>=1.7.0",
    "python-dotenv>=1.0.1",
    "google-api-python-client>=2.184.0",
    "polars>=1.8.2",
    "peft>=0.13.2",
    "jaxtyping>=0.2.19",
    "setuptools>=75.3.2",
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
